rerere: try_merge() should use LL_MERGE_ERROR when it means an error
When preimage or postimage files cannot be read, try_merge() function
returns LL_MERGE_CONFLICT.  To all of its callers, this does not make
them wrong things per-se, as they are only checking if the result is 0
and LL_MERGE_CONFLICT is not 0.  But it is an error if we fail to read
the input we expect to be able to read; return LL_MERGE_ERROR instead.

Signed-off-by: Junio C Hamano <gitster@pobox.com>