rerere: modernize use of empty strbuf
Back when the code to hash the contents in the strbuf "one" (and
"two") including its terminating NUL was written, a freshly
initialized strbuf had NULL in its .buf member, so it was an error
to say "update(one.buf, one.len + 1)", which was corrected by
b4833a2c (rerere: Fix use of an empty strbuf.buf, 2007-09-26).

But soon after that, b315c5c0 (strbuf change: be sure ->buf is never
ever NULL., 2007-09-27) introduced strbuf_slopbuf mechanism that
ensures that .buf member of a strbuf is *never* NULL.  A freshly
initialized and empty strbuf uses a static piece of memory that has
NUL in it, with its .len member set to 0, so we can always safely
use from offset 0 of .buf[] array for (one.len + 1) bytes.

Simplify the code by essentially reverting b4833a2c.

Signed-off-by: Junio C Hamano <gitster@pobox.com>