rerere: plug small strbuf leak
4af32207 (rerere: teach rerere to handle nested conflicts,
2018-08-05) introduced an extra strbuf to handle_conflict() function
to make it handle nested conflicts better.  After failing to handle
what appears to be an inner conflict (perhaps it was not a nested
conflicted region after all), this strbuf however was never released.

Narrow the scope of this strbuf to the small block that is about the
nested conflict, and make sure we release it when we are done.

Signed-off-by: Junio C Hamano <gitster@pobox.com>
---
cc: Thomas Gummerer <t.gummerer@gmail.com>